+++ blueprints/ember-fountainhead/index.js
@@ -1,8 +1,4 @@
 'use strict';
-const EOL = require('os').EOL;
-const readWatchmanConfig = require('../../lib/read-watchman-config.js');
-const updateWatchmanConfig = require('../../lib/update-watchman-config.js');
-
 /**
  * When an addon has an `index.js` file under `/blueprints/ADDON_NAME`, Ember
  * CLI will call the hooks defined there. We use this to handle anything that
@@ -12,10 +8,7 @@ const updateWatchmanConfig = require('../../lib/update-watchman-config.js');
  */
 
 /**
- * Fountainhead index blueprint is used to update the consuming application's
- * `.gitignore` to ignore the auto generated `vendor/feature-flags.js` file
- * created by `ember-radical`. This file is used to create global variables if
- * the consuming application has turned off JSUglify code stripping.
+ * Fountainhead's blueprint prints a friendly hello after install.
  * @class EmberFountainhead.Index
  * @constructor
  * @extends EmberCLI.Blueprint
@@ -26,26 +19,13 @@ module.exports = {
     // not specified (since that doesn't actually matter
     // to us
   },
-
-  // Add an ignore for generated vendor feature flags
   /**
-   * After install of addon read in consuming application's gitignore. If there
-   * isn't an entry for `vendor/feature-flags.js`, use node `fs` to append it.
+   * Say a quick hello.
    * @method afterInstall
-   * @return {undefined}
    */
   afterInstall() {
-    let watchmanConfig;
-
-    if (this.project.isEmberCLIAddon()) {
-      // Reads the project's .watchmanconfig file
-      watchmanConfig = readWatchmanConfig();
-      // Adds the public folder in a test dummy app to the ignore_dirs prop
-      // in the .watchmanconfig file to prevent infinite rerenders when docs
-      // are built on live reload.
-      updateWatchmanConfig(watchmanConfig);
-    }
-
-    console.log(`Thanks for installing Ember Fountainhead${EOL}You can run 'ember-docs' to generate you documentation`);
+    console.log('Thanks for installing Ember Fountainhead');
+    console.log('Run \'ember docs\' to generate you documentation files');
+    // this.project.isEmberCLIAddon()
   }
 };